Fix CancellationTokenMultiplexer return to pool
Fixes a operator precedence bug and adds tests.

If the PooledCancellationTokenSource had inlined tokens but no extraTokens, the Count would evaluate to 0 which would prevent the CancellationTokenRegistrations from being disposed.

The PooledCancellationTokenSource would then be returned to the pool but still be cancelable by the tokens which were preivously attached.